# Use `bounded` rather than `synchronous` in `parJoin`