@armanbilge 2022-10-27
Some related commits I found digging through the blame.
- https://github.com/typelevel/fs2/pull/2326/commits/749879c5e0fc755275a89afba79ebf23b1c86219
- https://github.com/typelevel/fs2/pull/2326/commits/1c0b491e3d8c5ab331dc65d2d3e05a37357b9f0a