@djspiewak 2022-10-27
Paul Chiusano in 2016. Wow. This is genuinely among the oldest lines of code remaining in the project.

What I find fascinating is, if you look in scalaz-stream, there's a predecessor to `parJoin` which has the same signature and functionality. I wrote that version back sometime in 2014, and it used queue with a bound of `maxOpen` to accept results. :-P So I think Paul's rewrite was the point at which it got intentionally adjusted. I wish he had written more about his rationale at the time.