@djspiewak 2022-10-27
In the interest of reporting negative results, I modified the test to `Blackhole` tunably in the upstreams or the downstreams. The hypothesis being that this change is most impactful when the upstreams are *faster* than the downstream. This corresponds to a flow bias of `downstream` in the following:

## Before

```
[info] Benchmark                    (concurrent)  (flowBias)   Mode  Cnt   Score   Error  Units
[info] ConcurrentBenchmark.parJoin             1    upstream  thrpt   10  28.166 ± 0.208  ops/s
[info] ConcurrentBenchmark.parJoin             1  downstream  thrpt   10  27.395 ± 0.410  ops/s
[info] ConcurrentBenchmark.parJoin            16    upstream  thrpt   10  45.193 ± 0.574  ops/s
[info] ConcurrentBenchmark.parJoin            16  downstream  thrpt   10  44.430 ± 0.984  ops/s
[info] ConcurrentBenchmark.parJoin           256    upstream  thrpt   10  44.172 ± 0.717  ops/s
[info] ConcurrentBenchmark.parJoin           256  downstream  thrpt   10  43.388 ± 0.575  ops/s
```

## After

```
[info] Benchmark                    (concurrent)  (flowBias)   Mode  Cnt   Score   Error  Units
[info] ConcurrentBenchmark.parJoin             1    upstream  thrpt   10  27.524 ± 0.362  ops/s
[info] ConcurrentBenchmark.parJoin             1  downstream  thrpt   10  29.582 ± 3.111  ops/s
[info] ConcurrentBenchmark.parJoin            16    upstream  thrpt   10  43.907 ± 0.796  ops/s
[info] ConcurrentBenchmark.parJoin            16  downstream  thrpt   10  43.086 ± 0.627  ops/s
[info] ConcurrentBenchmark.parJoin           256    upstream  thrpt   10  44.689 ± 0.792  ops/s
[info] ConcurrentBenchmark.parJoin           256  downstream  thrpt   10  44.322 ± 0.777  ops/s
```

## After + #2856

```
[info] Benchmark                    (concurrent)  (flowBias)   Mode  Cnt   Score   Error  Units
[info] ConcurrentBenchmark.parJoin             1    upstream  thrpt   10  26.651 ± 0.304  ops/s
[info] ConcurrentBenchmark.parJoin             1  downstream  thrpt   10  25.519 ± 0.201  ops/s
[info] ConcurrentBenchmark.parJoin            16    upstream  thrpt   10  44.321 ± 0.717  ops/s
[info] ConcurrentBenchmark.parJoin            16  downstream  thrpt   10  45.148 ± 0.575  ops/s
[info] ConcurrentBenchmark.parJoin           256    upstream  thrpt   10  45.709 ± 0.826  ops/s
[info] ConcurrentBenchmark.parJoin           256  downstream  thrpt   10  43.606 ± 0.634  ops/s
```

## Summary

It uh… doesn't matter? None of it? I think we're bounded by something else here and it's not intuitive to me what that might be.