@djspiewak 2022-11-19
Yeah I still feel like this is right on first principles, but I can't measure an improvement, so let's close this for now.